+++ src/fidesops/ops/api/deps.py
@@ -23,7 +23,7 @@ def get_db() -> Generator:
         )
 
     try:
-        db = _get_session()
+        db = get_api_session()
         yield db
     finally:
         db.close()
@@ -32,14 +32,14 @@ def get_db() -> Generator:
 def get_db_for_health_check() -> Generator:
     """Gets a database session regardless of whether the application db is disabled, for a health check."""
     try:
-        db = _get_session()
+        db = get_api_session()
         yield db
     finally:
         db.close()
 
 
-def _get_session() -> Session:
-    """Gets a database session"""
+def get_api_session() -> Session:
+    """Gets the shared database session to use for API functionality"""
     global _engine  # pylint: disable=W0603
     if not _engine:
         _engine = get_db_engine(config=config)