+++ src/fidesops/ops/service/connectors/saas/connector_registry_service.py
@@ -1,9 +1,12 @@
 from __future__ import annotations
 
+import logging
 from os.path import exists
-from typing import Dict, List, Optional
+from typing import Dict, Iterable, List, Optional, Union
 
 from fideslib.core.config import load_toml
+from packaging.version import LegacyVersion, Version
+from packaging.version import parse as parse_version
 from pydantic import BaseModel, validator
 from sqlalchemy.orm import Session
 
@@ -27,6 +30,8 @@ from fidesops.ops.util.saas_util import (
 
 _registry: Optional[ConnectorRegistry] = None
 registry_file = "data/saas/saas_connector_registry.toml"
+
+logger = logging.getLogger(__name__)
 
 
 class ConnectorTemplate(BaseModel):
@@ -106,13 +111,18 @@ def create_connection_config_from_template_no_save(
     return connection_config
 
 
-def create_dataset_config_from_template(
+def upsert_dataset_config_from_template(
     db: Session,
     connection_config: ConnectionConfig,
     template: ConnectorTemplate,
     template_values: SaasConnectionTemplateValues,
 ) -> DatasetConfig:
-    """Creates a DatasetConfig from a template and associates it with a ConnectionConfig"""
+    """
+    Creates a `DatasetConfig` from a template
+    and associates it with a ConnectionConfig.
+    If the `DatasetConfig` already exists in the db,
+    then the existing record is updated.
+    """
     # Load the dataset config from template and replace every instance of "<instance_fides_key>" with the fides_key
     # the user has chosen
     dataset_from_template: Dict = load_dataset_with_replacement(
@@ -123,7 +133,7 @@ def create_dataset_config_from_template(
         "fides_key": template_values.instance_key,
         "dataset": dataset_from_template,
     }
-    dataset_config = DatasetConfig.create(db, data=data)
+    dataset_config = DatasetConfig.create_or_update(db, data=data)
     return dataset_config
 
 
@@ -133,3 +143,81 @@ def load_registry(config_file: str) -> ConnectorRegistry:
     if _registry is None:
         _registry = ConnectorRegistry.parse_obj(load_toml([config_file]))
     return _registry
+
+
+def update_saas_configs(registry: ConnectorRegistry, db: Session) -> None:
+    """
+    Updates SaaS config instances currently in the DB if to the
+    corresponding template in the registry are found.
+
+    Effectively an "update script" for SaaS config instances,
+    to be run on server bootstrap.
+    """
+    for connector_type in registry.connector_types():
+        logger.debug(
+            "Determining if any updates are needed for connectors of type %s based on templates...",
+            connector_type,
+        )
+        template: ConnectorTemplate = registry.get_connector_template(  # type: ignore
+            connector_type
+        )
+        saas_config_template = SaaSConfig.parse_obj(load_config(template.config))
+        template_version: Union[LegacyVersion, Version] = parse_version(
+            saas_config_template.version
+        )
+
+        connection_configs: Iterable[ConnectionConfig] = ConnectionConfig.filter(
+            db=db,
+            conditions=(ConnectionConfig.saas_config["type"].astext == connector_type),
+        ).all()
+        for connection_config in connection_configs:
+            saas_config_instance = SaaSConfig.parse_obj(connection_config.saas_config)
+            if parse_version(saas_config_instance.version) < template_version:
+                logger.info(
+                    "Updating SaaS config instance '%s' of type '%s' as its version, %s, was found to be lower than the template version %s",
+                    saas_config_instance.fides_key,
+                    connector_type,
+                    saas_config_instance.version,
+                    template_version,
+                )
+                try:
+                    update_saas_instance(
+                        db,
+                        connection_config,
+                        template,
+                        saas_config_instance,
+                    )
+                except Exception:
+                    logger.error(
+                        "Encountered error attempting to update SaaS config instance %s",
+                        saas_config_instance.fides_key,
+                        exc_info=True,
+                    )
+
+
+def update_saas_instance(
+    db: Session,
+    connection_config: ConnectionConfig,
+    template: ConnectorTemplate,
+    saas_config_instance: SaaSConfig,
+) -> None:
+    """
+    Replace in the DB the existing SaaS instance configuration data
+    (SaaSConfig, DatasetConfig) associated with the given ConnectionConfig
+    with new instance configuration data based on the given ConnectorTemplate
+    """
+    template_vals = SaasConnectionTemplateValues(
+        name=connection_config.name,
+        key=connection_config.key,
+        description=connection_config.description,
+        secrets=connection_config.secrets,
+        instance_key=saas_config_instance.fides_key,
+    )
+
+    config_from_template: Dict = load_config_with_replacement(
+        template.config, "<instance_fides_key>", template_vals.instance_key
+    )
+
+    connection_config.update_saas_config(db, SaaSConfig(**config_from_template))
+
+    upsert_dataset_config_from_template(db, connection_config, template, template_vals)