+++ tests/ops/api/v1/endpoints/test_connection_template_endpoints.py
@@ -506,7 +506,7 @@ class TestInstantiateConnectionFromTemplate:
         }
 
     @mock.patch(
-        "fidesops.ops.api.v1.endpoints.saas_config_endpoints.create_dataset_config_from_template"
+        "fidesops.ops.api.v1.endpoints.saas_config_endpoints.upsert_dataset_config_from_template"
     )
     def test_dataset_config_saving_fails(
         self, mock_create_dataset, db, generate_auth_header, api_client, base_url