+++ CHANGELOG.md
@@ -48,6 +48,7 @@ The types of changes are:
 * Enable new manual webhooks in privacy request execution [#1285](https://github.com/ethyca/fidesops/pull/1285)
 * Added human readable label to ConnectionType endpoint [#1297](https://github.com/ethyca/fidesops/pull/1297)
 * Add table for consent (#1301)[https://github.com/ethyca/fidesops/pull/1301]
+* Utility to update SaaS config instances based on template updates [#1307](https://github.com/ethyca/fidesops/pull/1307)
 
 ### Docs